+++ src/Owin.Compression.Standard/AssemblyInfo.fs
@@ -5,13 +5,13 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("Owin.Compression.Standard")>]
 [<assembly: AssemblyProductAttribute("Owin.Compression")>]
 [<assembly: AssemblyDescriptionAttribute("Compression (Deflate / GZip) module for Microsoft OWIN Selfhost filesystem pipeline.")>]
-[<assembly: AssemblyVersionAttribute("1.0.29")>]
-[<assembly: AssemblyFileVersionAttribute("1.0.29")>]
+[<assembly: AssemblyVersionAttribute("1.0.30")>]
+[<assembly: AssemblyFileVersionAttribute("1.0.30")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Owin.Compression.Standard"
     let [<Literal>] AssemblyProduct = "Owin.Compression"
     let [<Literal>] AssemblyDescription = "Compression (Deflate / GZip) module for Microsoft OWIN Selfhost filesystem pipeline."
-    let [<Literal>] AssemblyVersion = "1.0.29"
-    let [<Literal>] AssemblyFileVersion = "1.0.29"
+    let [<Literal>] AssemblyVersion = "1.0.30"
+    let [<Literal>] AssemblyFileVersion = "1.0.30"